+++ awc/CHANGES.md
@@ -6,6 +6,10 @@
 
 * Re-export `actix_http::client::Connector`
 
+* Session wide headers
+
+* Session wide basic and bearer auth
+
 
 ## [0.1.0-alpha.1] - 2019-03-28